requirements.txt
asgiref==3.7.2
Django==4.2.4
django-admin-material-dashboard==1.0.10
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
phonenumbers==8.13.18
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
